detect: only write one label to avoid labels expansion (#160)
In the previous design, `detect` would attempt to discover all languages
and modules in the repo. If a repo contains many programming languages,
this would result in labels becoming very long and difficult to read.

Therefore, I modified this logic. Now, `detect` will only discover the
programming language with the most file entries. Additionally, if there
are modules present, it will no longer write the language label, as the
module is what we are more concerned about.